+++ example/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:hydrated_riverpod/hydrated_riverpod.dart';
 import 'package:path_provider/path_provider.dart';
+import 'package:riverpod_hydrated/hydrated_riverpod.dart';
 
 import 'pages/home_page.dart';
 import 'providers/theme_provider.dart';