+++ example/lib/providers/theme_provider.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:hydrated_riverpod/hydrated_riverpod.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:riverpod_hydrated/hydrated_riverpod.dart';
 
 /// Theme mode notifier - persists user's theme preference
 class ThemeModeNotifier extends HydratedNotifier<ThemeMode> {