+++ example/lib/providers/todo_provider.dart
@@ -1,5 +1,5 @@
-import 'package:hydrated_riverpod/hydrated_riverpod.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:riverpod_hydrated/hydrated_riverpod.dart';
 
 import '../models/todo.dart';